requirements.txt
# Core Django Requirements
Django>=3.2,<4.0
djangorestframework>=3.12.4,<4.0
django-filter>=21.1,<22.0
django-cors-headers>=3.10.0,<4.0
django-extensions>=3.1.5,<4.0

# Database
psycopg2-binary>=2.9.1,<3.0

# API Documentation
drf-yasg>=1.20.0,<2.0

# Authentication
django-oauth-toolkit>=1.5.0,<2.0
djangorestframework-simplejwt>=5.0.0,<6.0

# Background Tasks
celery>=5.1.2,<6.0
django-celery-beat>=2.2.1,<3.0
django-celery-results>=2.2.0,<3.0

# Caching
redis>=4.0.0,<5.0

# Machine Learning
numpy>=1.21.0,<2.0
pandas>=1.3.0,<2.0
scikit-learn>=1.0.0,<2.0
tensorflow>=2.7.0,<3.0  # Optional, for LSTM models

# Monitoring & Logging
sentry-sdk>=1.5.0,<2.0
django-prometheus>=2.2.0,<3.0

# Testing
pytest>=6.2.5,<7.0
pytest-django>=4.5.0,<5.0
factory-boy>=3.2.0,<4.0

# Development Tools
black>=21.12b0,<22.0
isort>=5.10.0,<6.0
flake8>=4.0.0,<5.0
mypy>=0.910,<1.0 